Split up cli into modules
