+++ cli/client.py
@@ -1,0 +1,7 @@
+import nucleus
+
+
+def compose_client():
+    # TODO: Use env var!
+    client = nucleus.NucleusClient("test_0b302578b4164aed9f2454a107cb1915")
+    return client
